+++ src/Examine.Core/Search/IExamineSpatialCircle.cs
@@ -1,0 +1,7 @@
+﻿namespace Examine.Search
+{
+    public interface IExamineSpatialCircle : IExamineSpatialShape
+    {
+        double Radius { get; }
+    }
+}
