+++ src/Examine.Core/Search/IExamineSpatialEmptyShape.cs
@@ -1,0 +1,7 @@
+﻿namespace Examine.Search
+{
+    public interface IExamineSpatialEmptyShape : IExamineSpatialShape
+    {
+
+    }
+}
