+++ src/Examine.Core/Search/IExamineSpatialShape.cs
@@ -11,28 +11,4 @@ namespace Examine.Search
         IExamineSpatialPoint Center { get; }
         bool IsEmpty { get; }
     }
-
-    public interface IExamineSpatialEmptyShape : IExamineSpatialShape
-    {
-
-    }
-
-    public interface IExamineSpatialRectangle : IExamineSpatialShape
-    {
-        /// <summary>The left edge of the X coordinate.</summary>
-        double MinX { get; }
-
-        /// <summary>The bottom edge of the Y coordinate.</summary>
-        double MinY { get; }
-
-        /// <summary>The right edge of the X coordinate.</summary>
-        double MaxX { get; }
-
-        /// <summary>The top edge of the Y coordinate.</summary>
-        double MaxY { get; }
-    }
-    public interface IExamineSpatialCircle : IExamineSpatialShape
-    {
-        double Radius { get; }
-    }
 }
