+++ src/Examine.Lucene/Search/LuceneSearchOptions.cs
@@ -2,6 +2,8 @@ using System;
 using System.Globalization;
 using Lucene.Net.Documents;
 using Lucene.Net.Search;
+using Lucene.Net.Spatial;
+using Spatial4n.Context;
 
 namespace Examine.Lucene.Search
 {
@@ -76,5 +78,15 @@ namespace Examine.Lucene.Search
         //     Sets the default Lucene.Net.Documents.DateTools.Resolution used for certain field
         //     when no Lucene.Net.Documents.DateTools.Resolution is defined for this field.
         public DateResolution? DateResolution { get; set; }
+
+        /// <summary>
+        /// Spatial Query Strategy
+        /// </summary>
+        public SpatialStrategy SpatialStrategy { get; set; }
+
+        /// <summary>
+        /// Spatial Query Context. Defaults to Geo Spatial
+        /// </summary>
+        public SpatialContext SpatialContext { get; set; } = SpatialContext.Geo;
     }
 }
